+++ app/controllers/users_controller.rb
@@ -3,13 +3,15 @@
 # Allows admins to manage user accounts.
 class UsersController < ApplicationController
   before_action :authenticate_user!
-  before_action :check_system_admin, except: [:settings, :update_settings]
+  before_action :check_system_admin, except: [:settings, :update_settings, :change_password]
   before_action :set_user, only: [:show, :edit, :update, :destroy]
   before_action :redirect_without_user, only: [:show, :edit, :update, :destroy]
 
-  def settings
-  end
+  # # GET /settings
+  # def settings
+  # end
 
+  # PATCH /settings
   def update_settings
     if current_user.update(user_params)
       redirect_to settings_path, notice: 'Settings successfully updated.'
@@ -18,14 +20,32 @@ class UsersController < ApplicationController
     end
   end
 
+  # PATCH /change_password
+  def change_password
+    if current_user.valid_password?(params[:user][:current_password])
+      if current_user.reset_password(params[:user][:password], params[:user][:password_confirmation])
+        bypass_sign_in current_user
+        redirect_to settings_path, notice: 'Your password has been changed.'
+      else
+        render :settings
+      end
+    else
+      current_user.errors.add(:current_password, 'is invalid')
+      render :settings
+    end
+  end
+
+  # GET /users
   def index
     @order = scrub_order(User, params[:order], 'users.current_sign_in_at desc')
     @users = User.current.search(params[:search]).order(@order).page(params[:page]).per(40)
   end
 
-  def edit
-  end
+  # # GET /users/1/edit
+  # def edit
+  # end
 
+  # PATCH /users/1
   def update
     if @user.update(user_params)
       redirect_to @user, notice: 'User was successfully updated.'